+++ src/test/java/racingcar/car/controller/CarGameManagerTest.java
@@ -1,4 +1,4 @@
-package racingcar.car.domain;
+package racingcar.car.controller;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalStateException;
